fake obj consolidation
