+++ a10_neutron_lbaas/tests/unit/v2/test_handler_lb.py
@@ -14,14 +14,14 @@
 
 import mock
 import test_base
-
+import fake_objs
 import a10_neutron_lbaas.a10_exceptions as a10_ex
 
 
 class TestLB(test_base.UnitTestBase):
 
     def test_create(self):
-        m = test_base.FakeLoadBalancer()
+        m = fake_objs.FakeLoadBalancer()
         self.a.lb.create(None, m)
         s = str(self.a.last_client.mock_calls)
         self.assertTrue('call.slb.virtual_server.create' in s)
@@ -52,7 +52,7 @@ class TestLB(test_base.UnitTestBase):
             v['api_version'] = api_ver
             v['default_virtual_server_vrid'] = default_vrid
 
-        lb = test_base.FakeLoadBalancer()
+        lb = fake_objs.FakeLoadBalancer()
         self.a.lb.create(None, lb)
 
         create = self.a.last_client.slb.virtual_server.create
@@ -92,7 +92,7 @@ class TestLB(test_base.UnitTestBase):
     #         self.assertTrue(str(2222+x) in s)
 
     def test_update_down(self):
-        m = test_base.FakeLoadBalancer()
+        m = fake_objs.FakeLoadBalancer()
         m.admin_state_up = False
         self.a.lb.update(None, m, m)
         s = str(self.a.last_client.mock_calls)
@@ -102,7 +102,7 @@ class TestLB(test_base.UnitTestBase):
         self.assertTrue('DOWN' in s)
 
     def test_delete(self):
-        m = test_base.FakeLoadBalancer()
+        m = fake_objs.FakeLoadBalancer()
         self.a.lb.delete(None, m)
         s = str(self.a.last_client.mock_calls)
         self.assertTrue('call.slb.virtual_server.delete' in s)
@@ -119,7 +119,7 @@ class TestLB(test_base.UnitTestBase):
             pass
 
     def test_stats(self):
-        test_lb = test_base.FakeLoadBalancer()
+        test_lb = fake_objs.FakeLoadBalancer()
         self.a.lb.stats(None, test_lb)
 
         self.print_mocks()
