+++ a10_neutron_lbaas/tests/unit/v2/test_handler_member.py
@@ -13,8 +13,12 @@
 #    under the License.
 
 import mock
+import time
 import test_base
+import fake_objs
+import logging
 
+LOG = logging.getLogger(__name__)
 
 def return_one(*args):
     return 1
@@ -33,27 +37,27 @@ class TestMembers(test_base.UnitTestBase):
         self.a.member.neutron.member_count = return_two
 
     def test_get_ip(self):
-        m = test_base.FakeMember(pool=mock.MagicMock())
+        m = fake_objs.FakeMember(pool=mock.MagicMock())
         self.a.member.neutron.member_get_ip(None, m, False)
         self.print_mocks()
         self.a.neutron.member_get_ip.assert_called_with(
             None, m, False)
 
     def test_get_name(self):
-        m = test_base.FakeMember(pool=mock.MagicMock())
+        m = fake_objs.FakeMember(pool=mock.MagicMock())
         z = self.a.member._get_name(m, '1.1.1.1')
         self.assertEqual(z, '_get-o_1_1_1_1_neutron')
 
     def test_count(self):
         self.a.member.neutron.member_count(
-            None, test_base.FakeMember(pool=mock.MagicMock()))
+            None, fake_objs.FakeMember(pool=mock.MagicMock()))
 
     def _test_create(self, admin_state_up=True, uuid_name=False):
         if uuid_name:
             old = self.a.config.get('member_name_use_uuid')
             self.a.config._config.member_name_use_uuid = True
 
-        m = test_base.FakeMember(admin_state_up=admin_state_up,
+        m = fake_objs.FakeMember(admin_state_up=admin_state_up,
                                  pool=mock.MagicMock())
         ip = self.a.member.neutron.member_get_ip(None, m, True)
         if uuid_name:
@@ -86,7 +90,7 @@ class TestMembers(test_base.UnitTestBase):
         self._test_create(admin_state_up=False)
 
     def test_update_down(self):
-        m = test_base.FakeMember(False, pool=mock.MagicMock())
+        m = fake_objs.FakeMember(False, pool=mock.MagicMock())
         ip = self.a.member.neutron.member_get_ip(None, m, True)
         name = self.a.member._get_name(m, ip)
         self.a.member.update(None, m, m)
@@ -96,7 +100,7 @@ class TestMembers(test_base.UnitTestBase):
             axapi_args={'member': {}})
 
     def test_delete(self):
-        m = test_base.FakeMember(False, pool=mock.MagicMock())
+        m = fake_objs.FakeMember(False, pool=mock.MagicMock())
         ip = self.a.member.neutron.member_get_ip(None, m, True)
 
         self.set_count_1()
@@ -105,7 +109,7 @@ class TestMembers(test_base.UnitTestBase):
         self.a.last_client.slb.server.delete(ip)
 
     def test_delete_count_gt_one(self):
-        m = test_base.FakeMember(False, pool=mock.MagicMock())
+        m = fake_objs.FakeMember(False, pool=mock.MagicMock())
         ip = self.a.member.neutron.member_get_ip(None, m, True)
         name = self.a.member._get_name(m, ip)
 
@@ -114,3 +118,11 @@ class TestMembers(test_base.UnitTestBase):
 
         self.a.last_client.slb.service_group.member.delete.assert_called_with(
             m.pool_id, name, m.protocol_port)
+
+    def test_updating_operating_status(self):
+        test_lb = fake_objs.FakeLoadBalancer()
+        self._test_create()
+        time.sleep(2)
+        self.print_mocks()
+        LOG.info("TESTING: {0}".format(self.a.openstack_driver.member.mock_calls))
+        s = str(self.a.last_client.mock_calls)
