+++ a10_neutron_lbaas/tests/unit/v2/test_handler_pool.py
@@ -14,7 +14,7 @@
 
 import mock
 import test_base
-
+import fake_objs
 import a10_neutron_lbaas.a10_exceptions as a10_ex
 
 
@@ -46,7 +46,7 @@ class TestPools(test_base.UnitTestBase):
                         self.a.reset_mocks()
                         saw_exception = False
 
-                        pool = test_base.FakePool(p, m, pers, listener)
+                        pool = fake_objs.FakePool(p, m, pers, listener)
                         try:
                             self.a.pool.create(None, pool)
                         except a10_ex.UnsupportedFeature as e:
@@ -85,8 +85,8 @@ class TestPools(test_base.UnitTestBase):
     def test_update(self):
         pers1 = None
         pers2 = None
-        old_pool = test_base.FakePool('TCP', 'LEAST_CONNECTIONS', pers1, True)
-        pool = test_base.FakePool('TCP', 'ROUND_ROBIN', pers2, True)
+        old_pool = fake_objs.FakePool('TCP', 'LEAST_CONNECTIONS', pers1, True)
+        pool = fake_objs.FakePool('TCP', 'ROUND_ROBIN', pers2, True)
         self.a.pool.update(None, pool, old_pool)
         self.print_mocks()
         self.a.last_client.slb.service_group.update.assert_called_with(
@@ -107,7 +107,7 @@ class TestPools(test_base.UnitTestBase):
                     for lst in listeners:
                         self.a.reset_mocks()
 
-                        pool = test_base.FakePool('TCP', 'ROUND_ROBIN',
+                        pool = fake_objs.FakePool('TCP', 'ROUND_ROBIN',
                                                   pers, lst,
                                                   members=m,
                                                   hm=hm)
@@ -128,7 +128,7 @@ class TestPools(test_base.UnitTestBase):
                                 assert_called_with(pool.id))
 
     def _test_stats(self):
-        pool = test_base.FakePool('TCP', 'ROUND_ROBIN', None, False)
+        pool = fake_objs.FakePool('TCP', 'ROUND_ROBIN', None, False)
         actual = self.a.pool.stats(None, pool)
         return pool, actual
